Renaming docs to API
