+++ example/src/routes.js
@@ -3,7 +3,7 @@ import { Route, IndexRoute } from 'react-router';
 import Base from './controllers/base';
 import Home from './controllers/home/home';
 import Examples from './controllers/examples/examples';
-import Docs from './controllers/docs/docs';
+import API from './controllers/api/api';
 import GettingStarted from './controllers/getting-started/getting-started';
 
 export default (
@@ -11,7 +11,7 @@ export default (
 		<IndexRoute component={Home} />
 		<Route path="home" component={Home} />
 		<Route path="examples" component={Examples} />
-		<Route path="docs" component={Docs} />
+		<Route path="api" component={API} />
 		<Route path="getting-started" component={GettingStarted} />
 	</Route>
 );
